chore(tests): Reorganise tests in better scoped files
